Algotrading/requirements.txt
Django>=3.0,<4.0
djangorestframework==3.12.2
psycopg2-binary>=2.8
bs4==0.0.1
cvxpy==1.1.11
cvxopt==1.2.6
nltk==3.5
numpy==1.20.1
pandas==1.2.2
praw==7.2.0
pyportfolioopt==1.4.1
PuLP==2.4
scikit-learn==0.24.1
yahoo-fin==0.8.7
yfinance==0.1.55